SearchLandscapeBook, Skeleton 파일 분리
