+++ src/components/Search/SearchLandscapeBook/MetaWrapper.ts
@@ -1,0 +1,13 @@
+import styled from '@emotion/styled';
+
+const SearchBookMetaWrapper = styled.div`
+   display: flex;
+   flex-direction: column;
+   margin-left: 16px;
+   > dl {
+    margin-bottom: 4px;
+  }
+   width: 100%;
+`;
+
+export default SearchBookMetaWrapper;
