+++ src/components/Skeleton/Bar.ts
@@ -1,0 +1,10 @@
+import styled from '@emotion/styled';
+
+const SkeletonBar = styled.div`
+  background: linear-gradient(327.23deg, #F8F9FB 1.42%, #F1F1F3 49.17%, #F8F9FB 100%);
+  width: 130px;
+  height: 20px;
+  margin-bottom: 8px;
+`;
+
+export default SkeletonBar;
