+++ src/pages/search/index.tsx
@@ -23,11 +23,8 @@ import sentry from 'src/utils/sentry';
 import { useEventTracker } from 'src/hooks/useEventTracker';
 
 import { RootState } from 'src/store/config';
-import {
-  SearchLandscapeBook,
-  SkeletonBar,
-} from 'src/components/Book/SearchLandscapeBook';
 import { Pagination } from 'src/components/Pagination/Pagination';
+import SearchLandscapeBook from 'src/components/Search/SearchLandscapeBook';
 import useIsTablet from 'src/hooks/useIsTablet';
 import { AdultExcludeToggle, FilterSelector } from 'src/components/Search';
 import { useRouter } from 'next/router';
@@ -36,6 +33,7 @@ import { useSearchQueries } from 'src/hooks/useSearchQueries';
 import { booksActions } from 'src/services/books';
 import { ITEM_PER_PAGE, MAX_PAGE, runSearch } from 'src/utils/search';
 import { Border } from 'src/components/Tabs/SearchCategoryTab';
+import SkeletonCategoryTab from 'src/components/Skeleton/CategoryTab';
 
 const SearchResultSection = styled.section`
   max-width: 952px;
@@ -214,19 +212,6 @@ const SuggestButton = styled.a`
   border-radius: 3px;
 `;
 
-const CategorySkeleton = styled.ul`
-  padding: 8px 0 9px 0;
-  display: flex;
-  align-items: center;
-  > * {
-    flex: none;
-    + * {
-      margin-left: 10px;
-    }
-  }
-  ${orBelow(BreakPoint.LG, 'margin-left: 16px;')}
-`;
-
 function SearchPage() {
   const dispatch = useDispatch();
   const { query, calculateUpdateQuery } = useSearchQueries();
@@ -358,11 +343,7 @@ function SearchPage() {
           />
         ) : (
           <>
-            <CategorySkeleton>
-              {
-                [...Array(8)].map((i, index) => (<SkeletonBar width="68px" key={index} />))
-              }
-            </CategorySkeleton>
+            <SkeletonCategoryTab />
             <Border color={slateGray10} />
           </>
         )}
